requirements.txt
flask
torch
transformers
Pillow